simplify decryption on worker
